+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.6
 
  - Drop support for 1.6
+ - Update to EffectLib 2.0 - all effect class names have changed!
  - Per-spell color and particle overrides
  - Add canTarget helper for Entity filtering
  - Fix bypass_confusion parameter